requirements.txt
fastapi>=0.110.0
uvicorn>=0.31.0
python-dotenv>=1.0.0
openai>=1.65.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langchain>=0.3.0
pydantic>=2.10.0
langgraph>=0.3.0
